Explicitly allow public and Opennet certificates for most downloads
Recently the certificate setup of some Opennet webservers was changed.
Thus the download of workarounds and patches was no longer working.

Now we start again to accept Opennet certificates as well as public
certificates. This will allow a shift of all remaining Opennet-CA-signed
webserver certificates, if we want to do this.