Replace "on_workaround_outdoor_channel_switch" symlink with copy
The older firmware versions (everything before 56bc47670c9225b) failed
to handle the new symlink automatically.  Thus the symlink prevented the
installation of the "outdoor_channel_switch" workaround.
Swapping the symlink (within the package and within the workarounds
directory) would not have solved the issue, since the packaging would
not convert this "external" symlink (pointing outside of the tree) into
the file content without requiring further changes.

Thus we keep a copy of this script for now.  The "workarounds" copy of
this file will hopefully become obsolete in a few years.