Ignore return value of "grep" when counting or filtering
"grep" returns with an non-zero exitcode in case of no output at all.
For example "echo | grep -c ." exits with an error.
In case of filtering and counting we can safely ignore the result of
grep.
The exitcode is implicitly ignored whenever it is not the last part of a
pipe.
Proper exitcode handling is explicitly desired whenever "-q" is used.

This fixes an issue for the "get_captive_portal_client_count" function
(used for a munin plugin).