Switch to a non-DFS channel if wireless scan fails
There seem to be problems (sometimes) with a wireless scan, if the
interface is configured as a master and runs on a DFS channel:

  https://forum.openwrt.org/t/cannot-scan-on-5g-on-tl-wr902ac/42863/2

Thus we temporarily switch to a non-DFS, if the scan failed and if the
interface is in "ap" mode and uses a DFS channel.

This should ensure, that the wireless scan works under almost all
conditions.

As a downside this may break the HTTP session between the browser and the
webserver, if the user connected via the IP address of the wireless
interface.

Closes: #217