Manually add main IP to loopback interface instead of mesh VPN interfaces
Hosts that lack a local interface with the device's main IP had problems
with routing via their main IP.  Thus we added the main IP in this case
to each VPN mesh interface.  But this caused problems with olsrd, since
it did not announce the mesh VPN address (10.2.0.0/16).  Thus outgoing
traffic to internal Opennet services failed on the reverse path.
We can avoid this problem by adding the main IP to the loopback
interface instead of the mesh VPN interfaces.
This is similar to the default setup for OLSR2.

Closes: #230