on_workaround_wifi_master_hangs: detect stuck TX queue
We recently noticed that AP1.52 lost the connection to its neighbour,
that was not detected early (before losing the association after a
timeout).  In this case the TX queue (instead of the RX queue) got
stuck.

The script now detects both situations separately and resets the
wireless interface if one of these conditions applies.