http_request: allow Opennet-CA and public certificates
Previously "http_request" trusted only public certificates.
Currently there are some Opennet services (e.g. the dev server for the
download of "on_workaround" scripts or the Opennet service discovery),
that are using Opennet-CA certificates.  This may likely change in the
future, thus we should prepare for this change and allow both public and
Opennet-CA certificates for HTTPS requests.