Comply with best practive and exec 'make 'clean' before build
"Deletes contents of the directories /bin and /build_dir. This doesn't remove the toolchain,
and it also avoids cleaning architectures/targets other than the one you have selected in
your .config. It is a good practice to do make clean before a build to ensure that no outdated
artefacts have been left from the previous builds. That may not be necessary always, but as a
general rule it helps to ensure quality builds. "
https://openwrt.org/docs/guide-developer/toolchain/use-buildsystem